LRCI-3892 Remove unnecessary Jobs component
